+++ nengo/networks/circularconvolution.py
@@ -184,9 +184,7 @@ class CircularConvolution(nengo.Network):
         invert_a=False,
         invert_b=False,
         input_magnitude=1.0,
-        # fmt: off
         **kwargs
-        # fmt: on
     ):
         if "net" in kwargs:
             raise ObsoleteError("The 'net' argument is no longer supported.")